리팩토링 및 코드 패턴 정리

+++ front/project_schedule/saga.ts
@@ -12,8 +12,8 @@ import {
   ProjectScheduleVO
 } from 'project_schedule/domain';
 import { projectScheduleApi } from 'project_schedule/api';
-import { dialogActions } from 'components/Dialog';
-import { ApiStatus } from 'components/DataFieldProps';
+import { dialogAction } from 'dialog/action';
+import { getErrorMessage } from 'type/Error';
 
 function* watchId() {
   while (true) {
@@ -43,19 +43,24 @@ function* watchFilter() {
 
 function* watchAdd() {
   while (true) {
-    const action = yield take(projectScheduleAction.add);
-    const { payload: params } = action;
-    console.log(action);
+    const { payload: params } = yield take(projectScheduleAction.add);
     try {
+      yield put(projectScheduleAction.requestAdd('request'));
       const { projectId } = yield select((root: RootState) => root.projectSchedule);
-      yield put(projectScheduleAction.requestAdd(ApiStatus.REQUEST));
+      if (!projectId) {
+        const message = '프로젝트가 선택되지 않았습니다.';
+        yield put(dialogAction.openError(message));
+        yield put(projectScheduleAction.requestAdd(message));
+        continue;
+      }
       yield call(projectScheduleApi.add, projectId, params);
-      yield put(projectScheduleAction.requestAdd(ApiStatus.DONE));
-      yield put(dialogActions.openAlert('저장하였습니다.'));
+      yield put(projectScheduleAction.requestAdd('done'));
+      yield put(dialogAction.openAlert('등록하였습니다.'));
     }
     catch (e) {
-      console.error(e);
-      yield put(projectScheduleAction.requestAdd(ApiStatus.FAIL));
+      const message = getErrorMessage(projectScheduleAction.add, e);
+      yield put(dialogAction.openError(message));
+      yield put(projectScheduleAction.requestAdd(message));
     }
   }
 }
@@ -64,31 +69,39 @@ function* watchUpdate() {
   while (true) {
     const { payload: params } = yield take(projectScheduleAction.update);
     try {
+      yield put(projectScheduleAction.requestUpdate('request'));
       yield call(projectScheduleApi.update, params);
-      yield put(dialogActions.openAlert('저장하였습니다.'));
+      yield put(projectScheduleAction.requestUpdate('done'));
+      yield put(dialogAction.openAlert('변경하였습니다.'));
     }
     catch (e) {
-      yield put(dialogActions.openAlert({
-        children: '저장에 실패하였습니다.',
-        status:   'error',
-      }));
+      const message = getErrorMessage(projectScheduleAction.update, e);
+      yield put(dialogAction.openError(message));
+      yield put(projectScheduleAction.requestUpdate(message));
     }
   }
 }
 
 function* watchDelete() {
   while (true) {
-    const { payload: id } = yield take(projectScheduleAction.deleteOne);
+    yield take(projectScheduleAction.deleteOne);
     try {
+      yield put(projectScheduleAction.requestDelete('request'));
+      const { id } = yield select((root: RootState) => root.projectSchedule);
+      if (!id) {
+        const message = '일정이 선택되지 않았습니다.';
+        yield put(dialogAction.openError(message));
+        yield put(projectScheduleAction.requestDelete(message));
+        continue;
+      }
       yield call(projectScheduleApi.deleteOne, id);
-      yield put(dialogActions.openAlert('삭제 했습니다.'));
+      yield put(projectScheduleAction.requestDelete('done'));
+      yield put(dialogAction.openAlert('삭제하였습니다.'));
     }
     catch (e) {
-      //TODO: 삭제 정책 후 수정 필요
-      yield put(dialogActions.openAlert({
-        children: '해당 일정은 삭제할 수 없습니다.',
-        status:   'error',
-      }));
+      const message = getErrorMessage(projectScheduleAction.deleteOne, e);
+      yield put(dialogAction.openError(message));
+      yield put(projectScheduleAction.requestDelete(message));
     }
   }
 }
